pyproject.toml
[project]
name = "legal-consultation-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "cryptography>=46.0.3",
    "django==5.0.1",
    "django-cors-headers==4.3.1",
    "djangorestframework==3.14.0",
    "drf-spectacular-sidecar>=2025.10.1",
    "drf-spectacular[sidecar]>=0.28.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "pymysql>=1.1.2",
    "python-dotenv==1.0.0",
    "redis>=7.0.0",
    "requests==2.31.0",
    "twilio==8.11.1",
]
